// web-client/src/shared/lib/types.ts
export type NewsArticleResponse = {
  id: number;
  title: string;
  slug: string;
  summary: string;
  content: string;
  status: "DRAFT" | "PUBLISHED" | "ARCHIVED" | "DELETED";
  category: string;
  tags: string[];
  likes: number;
  bookmarks: number;
  publishedAt: string;
  liked?: boolean;
  bookmarked?: boolean;
};

export type CommentResponse = {
  id: number;
  content: string;
  username: string;
  likes: number;
  liked: boolean;
  replies: CommentResponse[];
  publishedAt: string;
  parentId?: number;
};